requirements.txt
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
django-tinymce4-lite==1.8.0
djangorestframework==3.12.2
gunicorn==20.0.4
iniconfig==1.1.1
isort==5.6.4
jsmin==2.2.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.7
Pillow==8.1.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
